requirements.txt
torch==2.0.0
torchvision==0.15.1
numba==0.58.0
einops==0.6.1
gym==0.26.2
scipy==1.10.1
diffusers==0.21.2
h5py==3.9.0
h5py-cache==1.0
gin-config==0.5.0
absl-py==1.4.0
scikit-learn==1.3.0
